backend/requirements.txt
aiofiles==23.1.0
alembic==1.11.1
asyncpg==0.27.0
black==23.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
httpx==0.24.1
passlib==1.7.4
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.15
uvicorn==0.22.0
